MPS-34133 use language source to find out its runtimes
We can not wait for Language to get deployed to figure out it got a runtime dependency, have to look into its source.
However, to address the difference b/w SLanguage.getRuntimeModules(), deployed and transitive, and Language.getRuntimeModules(), source and restricted to module only, I decided to keep intersect, to make sure present dependency versions may stay intact.
There are numerous ways to address this, most appealing being not to keep dependency information at all (or at least not as part of the module descriptor). Moving runtimes to generator (it's a  'generation target', after all) is viable, too. For now, rely on the fact that we've got Language instance in Generator almost all the time.