http-support: expose netty fields
Expose underlying netty fields with getters in HttpRequest.
Sometimes the DSL and methods on HttpRequest are too limiting.

Use cases that require to set specific headers in the response field can
not do that at the moment because for that access to the FullHttpResponse
is required.

Use cases that want to implement a websocket handler can't do that because
it requires access to the channel and HttpRequest to do the handshake.

This change exposes these filed and allows experienced users to implement
http APIs similar to what is possible with plugins directly in intellij.