MPSSPRT-388 refactor log4j appender in generator
I'm uncertain whether log tracking makes a lot of sense, client code shall use high-level mechanism to convey messages to user.
However, given vast number of MPS subsystems that may log a message with a pointer to transient object, might be polite to try to keep these for navigation
Note, until we switch to a newer IDEA Platform (22.1), this code is not active. New platform is already on the way now, I don't expect this state to last.