MPSSPRT-388 No explicit log configuration in WorkerBase
The first thing worker does is initializing of IDEA/MPS environment, which is responsible to configure logging in a platform-specific way.
However, there's also EnvironmentConfig code that may want to log messages, so initialize log with the highest priority.
Need to come up with a better approach which would respect desired log level coming from Script.