Give descriptive names to library dependencies
It's not clear from "app.jar", "Util" or "3rd-party" what this dependency is about.
It's ok to get the same jar in classpath, it's not ok to spend time figuring out what a dependency is about.
With IDEA re-packaging its libraries, our attempt to use names of their libraris is futile.