MPSI-30 Ignore tests when they are commented out
This is a reasonable addition to the current test execution model.
We act like commented test nodes do not exist.

Next step will be providing a way to ignore a specific test or a group
of tests, for example via annotations or an additional flag.
Currently it seems to be excessive, since ignore annotations
in the text editor play much bigger role than in a projectional editor.
Commented nodes still contain static references to the nodes around
them.