DynamicReference: revert MPS-19497 hack
MPS shall not ask for a target of a dynamic reference which is not part of a model. It does, however, at least for pasted nodes (see CopyPasteUtil.processReferencesOut() and MPS-33942). I don't think the hack is the right way to go, now we tell null as we can not resolve node in such scenario, and, instead, rely on processReferencesOut() code to grab resolve info (oldTargetNode == null)) to establish a new dynamic reference that could get resolved later).
I'm about to convert DynamicReference to AssociationData religion, and don't need extra state to care about.