MPSI-43 Project-Modules are not shown when added via git-pull in MPS
the doc in the commit describes the solution,
 but in short .mps/modules.xml updates are ignored right after the startup since the directory is not loaded, so vfs cannot recognize new changes in modules.xml.
To be honest I did not check how intellij loads this directory into the snapshot (because eventually it does load it for sure), and probably it is worth exploring.
Fsnotifier works correctly, giving out all dirty paths whenever they are changed on disk.

Idea does not have such a problem since it is file-based
and there are many ways for .idea to fall into vfs snapshot.
Even when after first alt-tab it does not recognize a new module, second
alt-tab always does the job.

Proper solution would be to push a similar logic into intellij repo.