MPS-34910 EditorComponentTrackService to move notifications out of EC
Separate EditorComponent, Swing/UI element, from IDEA/Platform tracking and notification.
EC to be responsible for UI stuff, and external code tracks changes in the platform that may affect UI and bear responsibility to update visuals as appropriate (if and when necessary).
One of the platform change is reload of languages that require rebuild of an editor.