MPS-34680 / Kotlin: replace intermediate type by new type attribute
Intermediate types allowed to recover additional types on top of the node's regular type (eg. on a lambda, we can get lambda type, type of 'it' inside lambda and type of 'this' inside lambda).
This approach has issues since it was querying the typesystem aside of the regular checking process, creating computation duplication and slowness.
The new approach is to add attributes to the returned type with TypeAdditionalData. This do not require any change to the existing API (see changes in InferredTypeReference) and will increase performances when using lambdas and for loops.