build.testManifest no longer use deprecated smodel expressions
I decided to go with ModuleIdentity instead of LanguageIdentity as the focus there is
to include project languages. As long as they are not necessarily deployed, I see no reason
to stick to SLanguage. Rather, have to check this with a constraint/checking rule (if necessary).
Decided not to add migration as the language deemed for internal use, afaiu, and there are no uss in MPS to migrate.