Collection runtime: deprecated toListSequence, replace with toList
ToListSequence() was previously implemented as either a cast (listSequence.toList => return this) or a copy operation (new list(this))
Its only usage was the toList operator in collection language, which was documented as a copy operator (https://www.jetbrains.com/help/mps/sequence.html#tolist)
Keeping the orginal list goes against the intent of the developer (toList is used to explicitely create a new list), and could potentially create unwanted side effects.
For instance:
| public list<int> internalState = new arraylist<int>;
| public sequence<int> getSequence()  {  return internalState; }
| public void fn(s)  {
|   test.getSequence().toList.add(1);
|
|   // prints 1, should print 0
|   System.out.println(test.getSequence().size);
| }

In addition, this fixes MPS-14004 generation issue: toList now actually returns a list, as the typesystem assert.
For that specific issue, another option would be to use collectors (which seems to provide more flexibility wrt to type checking on java side), downsides would be increased generation size (collect(MpsCollectors.toListSequence())) but the collector pattern would be a good addition anyway.

Another option could be to make this copy lazy (copy to a new list only when someone tries to alter the list), but that would be harder to maintain and prone to new issues (and there is no guarantee this feature is widely used and impacts performances somewhere).