MPS-27879 lang.build: confusing error when macro doesn't have default value
I don't quite understand the logic when we first expand paths with macro
just to shrink them later in PathConverter. To me, looks like we can do it the moment
we need lang.build paths, and keep original values. There's no other code that uses
these except ModuleChecker.
The check, if necessary, has to be part of ModuleChecker anyway, as it's odd to fail to
load a module with a variable, but to tolerate identical, hand-coded value in a build project.