MPS-27879 lang.build: match path spec in build project and module descriptor
whether module descriptor uses platform_lib or mps_home, we have to stick to the same in build project.
Can't rely on the fact macro default values in build project point to the proper/actual location.
This is a follow-up change to earlier commit when we stop translating macro values back and forth,
keeping original macro name intact. One nice side effect is that when a migration erroneously changes
$mps_home/lib to $platform_lib, the change won't go unnoticed, as there are build errors then.