Introduce overriding typechecking session
A generator session may request a new typechecking session on a thread
that is running other typechecking-related computations that also require
a typechecking session, which is irrelevant to the generator.
In this case the generator takes priority and the session requested by it
takes precedence, with the existing one being "overridden" and restored later.