Fix race condition
Condition variables are used incorrectly in this code.

If signalInitialized() is called before waiting for it, it will not be
registered and the thread unnecessarily waits for 100 seconds.

In addition, if the timeout expires we probably want to report this as
an error but the thread instead continues to run.

Using a CountDownLatch instead of Condition and ReentrantLock fixes both
problems.