lang.editor: boilerplate code into DefaultSubstituteMenuItem rt class
There's no reason to have the same code again and again in each DefaultSubstituteMenuItem subclass we generate.
Instead, keep it in runtime, yet keep switches for possible extensions.
For MPS, this gives greater control over default implementation (what and how we access facilities to provide an answer) and limits exposure of internal utility classes.
Have to keep _context variable as it's a mechanism editor generator resolves some of its ConceptFunctionParameters