Drop check for isChanged() before repository save; MPS-36512
The flag `EditableSModule.isChanged()` isn't affected by changes made
to models as a result of editing. In a normal editing scenario none of the project
modules are going to have this flag set to true, thus Save All command
is going to have no effect, which results in all edits being lost in case
the app is exited.