Runtime/test modules don't provide extensions to MPS
ext=yes due to migration of solutionKind=PLUGIN_XXX, which was introduced back in 2012 to help classloading from solutions.
See commits 5457ec49, ce180004 and 6d4100b0.
Likely, "plugin solution" at the time was supposed to mean a "solution to plug into MPS" for whatever purpose, including classloading (compile in MPS was not enough to tell if it's a code to try to load classes for MPS purposed)