migration: replace IDEA.ProjectComponent with dynamic CoreComponent
to keep setXXX "configuration" mechanism as is, try to replace ProjectComponent not with a project service, but with a dynamic CC.
With that, can eventually move the code to MPS space (service would need to stay in IDEA's), just need to move NoPendingMigrationTest to Test Binaries (mps module instead of testbench class)

IMPORTANT: to bundle migration.component module into migration plugin of MPS-as-IDEA plugin, decided to build a copy of j.m.migration.component module rather than share one from mpsMigrationCore. IdeaPlugin build runs separately, not a big deal to build the module once again. Using mpsMigrationCore approach breaks dependent BuildProjects (like mpsTestModules) as they fail to detect the module is part of plugin and just reference it by module jar (which is wrong but I don't want to deal with this buildlang defect as I plan to get rid of mpsMigrationCore separation altogether once j.m.migration.component becomes mps-managed).

FTR, MPS-as-IDEA plugin uses DynamicComponentWarden to register code that setXXX() to make sure it happens when appropriate (ModuleActivator code execution is not aligned with IDEA's startup sequence). For Big MPS it's not an issue as .workbench module depends on .component and there's confidence ModuleActivator has been executed once project parts get control