MPS-36867 associations in new nodes don't get proper resolve info
With SNode.getName() focused on persisted values, not constraints-derived (3887b1ca), we need to use [kernel]-generated code to access resolve info value. Although the use of SNodeOperation is not quite impressive, using Scope to figure out resolve info is not feasible now/here.
I refrained from moving IR into [kernel] (as it seems to denounce need for [smodel]/[kernel] separation), now I've come to terms, treating IR as an implementation detail on top of plain, straightforward SNode implementation. Even once/if SNode is moved to [smodel], IR is not part of anything exposed in [openapi] and, therefore, doesn't need to live in [smodel]