MPS-36688 respect change of module's classloading state
when a module is no longer capable to provide classes, we can't throw it away immediatelly as we need to unload its dependencies as well.
present implementation requires module to stay in the dependency graph. The fix keeps the module in the graph but marks it as "broken"
so that not the module nor its dependencies get re-loaded.