BaseLanguage: fix subtyping equality between classifier types
Previously, :==: was used, which does not take some rules into account (int :==: Integer as type parameter)

This reopens MPS-15014, as it allows List<List> to be assigned to List<List<Integer>> while forbidden in java. Having DSL types properly supported instead of having raw type properly supported seems to be a better option. Alternatively, we could spend time to bring proper support for this use case, or add warning about raw types in MPS (to at least show some errors that could point to the problem).
The previous implementation of the check (left and right side are classifier type) is insufficient to safely use :==:, the check should rather be "all descendant types should be classifier types" (but then it could be bypassed by using a equivalent type, eg. node<> instead of SNode).