CopyPaste: tell cut/copied NodeData and preserve UO accordingly
The idea is to copy SNode's user objects only when it's a set of unique nodes in the clipboard.
As long as we keep SNode's UID as UO, we shall preserve these on first Cut-Paste, but do not preserve on subsequent Paste.
For Copy-Paste, first Paste shall produce new UID right away, hence Copy action mimics first paste with PND.consume()