tests: extract junit5 libraries and stub into dedicated plugin
As long as 'unitTest.launcher' code sits a jar under plugin/lib (with Plugin CL), exposed as "provied" classes to MPS world but sometimes executed as a plain Java
(e.g. for out-of-process, bare JUnit5 launches), need to make sure junit5 classes could get available to launcher code both in a command-line as well as 'MPS' scenarios
(out-of-process with platform or MPS in-process).