MPS-36475: move mps-vcs modules out of mpsVcsPlatform into mpsVcs and mpsIdeaVcs
Build language, in `ModulePlugins`, considered these module were not part of a plugin because of the use of `project2plugins.get(module.containingRoot)` and the root of the module node (defined in mpsVcsPlatform) not containing the plugin (declared in mpsVcs).

This resulted in invalid generation of `<library>` elements under the `generate` task, which in turn led to classloading issues.