MPS-37229 model actions don't tolerate IDEA cancellation
WMA indicates all model actions as non-cancellable.
This approach isn't perfect as we need to incorporate our CancellableReadActions and/or provide a mechanism for a user to run cancellable code (if user is ok with that). However, at the moment, we have a lot of actions that end up in IDEA's underlying code (like VFS), and no MPS code is designed to deal with read cancellations (especially as a RuntimeException), so here we go with every model read/write/command denoted as non-cancellable.