MPS-37429: make MultiLambdaParameter and FlexibleBlock inference parts
These were previously determined to be an inference root, resulting in invalid local type.

While FlexibleBlock could be assumed not to be an expression in first hand (it holds statements), it is used in practice as part of expressions (If, When), which could require the whole enclosing statement to be inferred. This has been reflected on the concept description.