MPS-34747 Generator module disappear from a repo
To update an owned Generator, Language tries to find existing one matching by both name and id. Name doesn't match for a renamed generator, therefore the "old" Generator is treated stale, and the "new" one is attempted to get registered. As SRepository matches by ID only, the old instance is reused, just linked to the "new" (in fact, the same) owner Language instance. Then, stale Generator is removed, and an attempt to get its instance from a repository fails.