ModelValidationAdapter into MPS-managed code
To move ModelCheckerTool into MPS world and use lang.plugin.ToolDeclaration for it, we need to have code which accesses the tool in MPS space as well.
MVA accesses ModelCheckerTool but use of IDEA extpoint forces its implementation to reside in IDEA-managed code. With modern ModuleRuntime, there's no
such limitation.