MPS-37926: Modifying IdeaEnvironment to wait for the end on index build process after openning the project.
Some our tests may need IDEA indexes to be prepared before the test execution.
At the same time the indexing process is executed in the background and there is no guarantee that the indexing
process(es) will be finished before the actual test run.

Any platform write (command) executed by the test will block running background indexing process, so
there is a chance that the index will be ready only in the end of test exectution or aven later..

See: https://youtrack.jetbrains.com/issue/MPS-37926/Indices-not-built-properly-in-IdeaEnvironment