MPS-38365 generation with transients and inplace fails
With introduction of LocalNodePtr for same-model association and removal of SModel.changeModelReference(),
cloning of transient model leads to same-model associations being redirected to original nodes (known peculiarity of
generator custom clone code, see FIXME in CloneUtil.clone()). VariableReference under ClosureLiteral pointed to
nodes outside the closure (in fact, inside the closure, but in a different (previous) transient model), leading to
generation of erroneous code.

Not sure what's role of in-place transformation here, though.

Besides, as an effort to handle these scenarios in general, convert Indirect to Local in makeIndirect(). Doesn't help M2M scenarios as models are detached (not in repo), and we don't makeDirect/makeIndirect for free-floating nodes