Deprecate AM -> FileSystem exposure
With MPS (and derivatives) code heavily invested into AbstractModule being the basic/ubiquitous implementation and the need for other file-less and MD-less module implementations (e.g. in-memory modules that allow e.g. dependency manipulation), the effort is to make AM bound to MD only, with optional IFile gradually moving outside of AM into a dedicated subclass (yet to be introduced)