TempModels: refactor single use of fs-backed, compiled, reloaded temp module
Console is the only client of the forDefaultModuleWithSourceAndClassesGen() method, and, in fact, the only user of NaiveJavaFacet with temp locations.
Perhaps, eventually the temp model would drop ModuleDescriptor as well, as the reason for that is likely model root descriptors, that are not in use any more.
Indeed, there are still some assumptions about temp models being AbstractModule, we need to find and eliminate these.