Notify if no output location found (no generation facet); MPS-39138
If a module does not have any generation facets activated,
make may not save any files at the end of generation.
This is only reported as a warning in the messages view
and can easily be overlooked.

This conditinon now triggers a notification in the IDE
that clearly states the problem and is not easy to ignore.