Ensure make session is finished before launching tests
Sporadic test failure is due to a race with make session, which in turn
forces "dumb mode" of file indexer when saving generated contents.
It's because of indexer's "dumb mode" a request to find usages fails,
which breaks refactoring logic.

This change is in the hope that, despite racing with make session, a call
to MakeServiceComponent#isSessionActive() is sensible.