Delegate NewElement to Idea action outside of Logical View; MPS-39260
Apparently a keyboard shortcuts conflict causes Idea's default
NewElement action to be activated. Both MPS's own NewElement and
the default one are applicable in the same context, so the conflict
is unavoidable.

The solution appears to be to delegate to the default action if MPS's
NewElement is invoked outside of its usual context, that is the project pane.