MigrationSession: drop AppliedScript.asLegacy(), 4/4
The idea behind PostponedState is quite confusing, as in most cases we could just
use current() and go from there. However, decided not to change it too much.
With this change, if new modules require same migrations that others, previously
ignored, we don't recognize the change as 'migration required'. To me, it seems fine
as if I didn't migrate other modules, why would I want to trigger migrations when I
add more uses of old stuff?